AOC 2022: Corrected few old code

+++ aoc_2022/day6p2/inc/AocUtils.h
@@ -1,10 +1,9 @@
 /* Copyright [2022-2023] Balamurugan R<emailstorbala@gmail.com> */
 #pragma once
 #include <string>
-#include <string_view>
 
 using std::string;
 
 struct AocUtils {
-  int GetMarkerCharacter(const string &inpStr);
+  size_t GetMarkerCharacter(const string &inpStr);
 };
